Add APIs for specifying exact handle locations in scripts; rename some APIs, fix bugs in dragUpdateFunc implementations. Fix compilation errors in some test playgrounds.
Some of the implementations of dragUpdateFunc assumed that the point argument was for the end location, but it’s actually the start location of the drag update. The end location can be computed by applying the vector in the second argument.
The renamed APIs needed corrections to be consistent with other APIs (hintX instead of X when X can change as a result of dragging).